+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Version 1.1.0
+Miscellaneous updates
+
+* Support for creating S3 bucket and assigning policy
+* Support for defining load balancer within a CreateLoadBalancer task
+* Extracted describe and delete tasks 
+
 # Version 1.0.0
 Initial public release